notebook and discsampler left to fix
