+++ graphlearn/utils/draw.py
@@ -1,9 +1,14 @@
 import pylab as plt
 from eden.util.display import draw_graph
+from eden.graph import Vectorizer
+import networkx as nx
 
-from myeden import *
-
-
+def expand_edges(graph):
+    '''
+    convenience wrapper
+    '''
+    vectorizer = Vectorizer(complexity= 3)
+    return vectorizer._edge_to_vertex_transform(graph)
 '''
         
     functions to draw graphs:
@@ -15,6 +20,13 @@ from myeden import *
             drawing a list of graphs
 '''
 
+
+def expand_edges(graph):
+    '''
+    convenience wrapper
+    '''
+    vectorizer = Vectorizer(complexity= 3)
+    return vectorizer._edge_to_vertex_transform(graph)
 def graph_clean(graph):
 
     '''
@@ -87,6 +99,21 @@ def draw_grammar_stats(grammar):
     plt.show()
 
 
+from collections import defaultdict
+def calc_stats_from_grammar(grammar):
+    count_corehashes = defaultdict(int)
+    count_interfacehashes = defaultdict(int)
+    corecounter = defaultdict(int)
+    intercounter = defaultdict(int)
+    for ih in grammar.keys():
+        for ch in grammar[ih].keys():
+            # go over all the combos
+            count_corehashes[ch]+=1
+            count_interfacehashes[ih]+=1
+            count= grammar[ih][ch].count
+            corecounter[ch]+=count
+            intercounter[ih]+=count
+    return count_corehashes,count_interfacehashes,corecounter,intercounter
 
 def display(G, size=6, font_size=15, node_size=200, node_border=False, delabeledges=True, contract=False,
             vertex_label='label'):
@@ -321,5 +348,31 @@ def row_draw_graph(graph,
         plt.title(title)
         #plt.show()
 
-
+def contract_edges(original_graph):
+    """
+        stealing from eden...
+        because i draw cores and interfaces there may be edge-nodes
+        that have no partner, eden gives error in this case.
+        i still want to see them :)
+    """
+    # start from a copy of the original graph
+    G = nx.Graph(original_graph)
+    # re-wire the endpoints of edge-vertices
+    for n, d in original_graph.nodes_iter(data=True):
+        if d.get('edge', False) == True:
+            # extract the endpoints
+            endpoints = [u for u in original_graph.neighbors(n)]
+            # assert (len(endpoints) == 2), 'ERROR: more than 2 endpoints'
+            if len(endpoints) != 2:
+                continue
+            u = endpoints[0]
+            v = endpoints[1]
+            # add the corresponding edge
+            G.add_edge(u, v, d)
+            # remove the edge-vertex
+            G.remove_node(n)
+        if d.get('node', False) == True:
+            # remove stale information
+            G.node[n].pop('remote_neighbours', None)
+    return G
